Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -57,7 +57,7 @@ def get_average_age_for_season(data, season)
   data.each do |keys, values|
     if keys == season
       values.each do |values|
-        ages.push(values["age"].to_i)
+        ages.push(values["age"].to_f)
       end
       end
     end
